drivers/cmsis-dap: update for newest protocol version
The capabilities INFO command can now return two bytes, without this
patch, the capabilities would simply not be read and left as 0 (i.e. no
capabilities).

cf. https://arm-software.github.io/CMSIS_5/DAP/html/group__DAP__Info.html ;
https://github.com/ARM-software/CMSIS_5/blob/116866fd74756c88096e37cbd0066fadad583cad/CMSIS/DAP/Firmware/Source/DAP.c#L100-L111

Reviewed-on: http://openocd.zylin.com/6373
Tested-by: jenkins

+++ src/jtag/drivers/cmsis_dap.h
@@ -16,7 +16,7 @@ struct cmsis_dap {
 	uint16_t packet_buffer_size;
 	uint8_t *command;
 	uint8_t *response;
-	uint8_t caps;
+	uint16_t caps;
 	uint8_t mode;
 	uint32_t swo_buf_sz;
 	bool trace_enabled;
